+++ markwon-core/src/main/java/io/noties/markwon/Markwon.java
@@ -192,6 +192,17 @@ public abstract class Markwon {
         @NonNull
         Builder usePlugins(@NonNull Iterable<? extends MarkwonPlugin> plugins);
 
+        /**
+         * Control if small chunks of non-finished markdown sentences (for example, a single `*` character)
+         * should be displayed/rendered as raw input instead of an empty string.
+         * <p>
+         * Since $nap; {@code true} by default, versions prior - {@code false}
+         *
+         * @since $nap;
+         */
+        @NonNull
+        Builder fallbackToRawInputWhenEmpty(boolean fallbackToRawInputWhenEmpty);
+
         @NonNull
         Markwon build();
     }